Revert solanum backport fixes
Reverts all 13 commits from the solanum-ircd backport series.

+++ modules/m_stats.c
@@ -70,7 +70,7 @@ mapi_hlist_av1 stats_hlist[] = {
 
 DECLARE_MODULE_AV2(stats, NULL, NULL, stats_clist, stats_hlist, NULL, NULL, NULL, stats_desc);
 
-const char *Lformat = "%s %d %"PRIu32" %"PRIu32" %"PRIu32" %"PRIu32" :%"PRId64" %"PRId64" %s";
+const char *Lformat = "%s %u %u %u %u %u :%u %u %s";
 
 static void stats_l_list(struct Client *s, const char *, bool, bool, rb_dlink_list *, char,
 				bool (*check_fn)(struct Client *source_p, struct Client *target_p));
@@ -1480,7 +1480,7 @@ stats_memory (struct Client *source_p)
 static void
 stats_servlinks (struct Client *source_p)
 {
-	static char Sformat[] = ":%s %d %s %s %d %"PRIu32" %"PRIu32" %"PRIu32" %"PRIu32" :%"PRId64" %"PRId64" %s";
+	static char Sformat[] = ":%s %d %s %s %u %u %u %u %u :%u %u %s";
 	long uptime, sendK, receiveK;
 	struct Client *target_p;
 	rb_dlink_node *ptr;
@@ -1509,13 +1509,13 @@ stats_servlinks (struct Client *source_p)
 			get_id(&me, source_p), RPL_STATSLINKINFO, get_id(source_p, source_p),
 			target_p->name,
 			(int) rb_linebuf_len (&target_p->localClient->buf_sendq),
-			target_p->localClient->sendM,
-			target_p->localClient->sendK,
-			target_p->localClient->receiveM,
-			target_p->localClient->receiveK,
-			(int64_t)(rb_current_time() - target_p->localClient->firsttime),
-			(int64_t)((rb_current_time() > target_p->localClient->lasttime) ?
-			 (rb_current_time() - target_p->localClient->lasttime) : 0),
+			(int) target_p->localClient->sendM,
+			(int) target_p->localClient->sendK,
+			(int) target_p->localClient->receiveM,
+			(int) target_p->localClient->receiveK,
+			rb_current_time() - target_p->localClient->firsttime,
+			(rb_current_time() > target_p->localClient->lasttime) ?
+			 (rb_current_time() - target_p->localClient->lasttime) : 0,
 			IsOperGeneral (source_p) ? show_capabilities (target_p) : "TS");
 	}
 
@@ -1669,13 +1669,13 @@ stats_l_client(struct Client *source_p, struct Client *target_p,
 		sendto_one_numeric(source_p, RPL_STATSLINKINFO, Lformat,
 				target_p->name,
 				(int) rb_linebuf_len(&target_p->localClient->buf_sendq),
-				target_p->localClient->sendM,
-				target_p->localClient->sendK,
-				target_p->localClient->receiveM,
-				target_p->localClient->receiveK,
-				(int64_t)(rb_current_time() - target_p->localClient->firsttime),
-				(int64_t)((rb_current_time() > target_p->localClient->lasttime) ?
-				 (rb_current_time() - target_p->localClient->lasttime) : 0),
+				(int) target_p->localClient->sendM,
+				(int) target_p->localClient->sendK,
+				(int) target_p->localClient->receiveM,
+				(int) target_p->localClient->receiveK,
+				rb_current_time() - target_p->localClient->firsttime,
+				(rb_current_time() > target_p->localClient->lasttime) ?
+				 (rb_current_time() - target_p->localClient->lasttime) : 0,
 				IsOperGeneral(source_p) ? show_capabilities(target_p) : "-");
 	}
 
@@ -1688,13 +1688,13 @@ stats_l_client(struct Client *source_p, struct Client *target_p,
 				     get_client_name(target_p, HIDE_IP)) :
 				    get_client_name(target_p, MASK_IP),
 				    (int) rb_linebuf_len(&target_p->localClient->buf_sendq),
-				    target_p->localClient->sendM,
-				    target_p->localClient->sendK,
-				    target_p->localClient->receiveM,
-				    target_p->localClient->receiveK,
-				    (int64_t)(rb_current_time() - target_p->localClient->firsttime),
-				    (int64_t)((rb_current_time() > target_p->localClient->lasttime) ?
-				     (rb_current_time() - target_p->localClient->lasttime) : 0),
+				    (int) target_p->localClient->sendM,
+				    (int) target_p->localClient->sendK,
+				    (int) target_p->localClient->receiveM,
+				    (int) target_p->localClient->receiveK,
+				    rb_current_time() - target_p->localClient->firsttime,
+				    (rb_current_time() > target_p->localClient->lasttime) ?
+				     (rb_current_time() - target_p->localClient->lasttime) : 0,
 				    "-");
 	}
 }
